changed dimensions to floats for convenience

+++ TI/Button.pde
@@ -1,9 +1,9 @@
 class Button{
   private String identity; //to identify what the button does
   private String identityA; // identity assigned by shuffle in Annoying Mode
-  public int x, y, wid, hei; //location and dimensions
+  public float x, y, wid, hei; //location and dimensions
   
-  public Button(String Identity, int X, int Y, int w, int h){
+  public Button(String Identity, float X, float Y, float w, float h){
     identity = Identity;
     identityA = Identity;
     x = X;
